pyproject.toml
[project]
name = "link-up"
version = "0.1.0"
description = "My project"
authors = [
    {name = "leonidig",email = "leonid200955@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
